+++ public/locales/en/common.json
@@ -26,7 +26,7 @@
   "loading": "Loading...",
   "loadingFinished": "Loading finished",
   "navigation": {
-    "backToEnrolmentForm": "Back to the registration form",
+    "backToSignupGroupForm": "Back to the registration form",
     "languages": {
       "en": "In English",
       "fi": "Suomeksi",
