+++ public/locales/fi/common.json
@@ -26,7 +26,7 @@
   "loading": "Ladataan...",
   "loadingFinished": "Lataaminen päättynyt",
   "navigation": {
-    "backToEnrolmentForm": "Palaa ilmoittautumiskaavakkeeseen",
+    "backToSignupGroupForm": "Palaa ilmoittautumiskaavakkeeseen",
     "languages": {
       "en": "In English",
       "fi": "Suomeksi",
