+++ src/domain/app/routes/constants.ts
@@ -1,7 +1,7 @@
 export enum ROUTES {
   CALLBACK = '/api/auth/callback/tunnistamo',
-  CREATE_ENROLMENT = '/registration/[registrationId]/enrolment/create',
-  CREATE_ENROLMENT_SUMMARY = '/registration/[registrationId]/enrolment/create/summary',
+  CREATE_SIGNUP_GROUP = '/registration/[registrationId]/signup-group/create',
+  CREATE_SIGNUP_GROUP_SUMMARY = '/registration/[registrationId]/signup-group/create/summary',
   EDIT_ENROLMENT = '/registration/[registrationId]/enrolment/[enrolmentId]/edit',
   ENROLMENT_COMPLETED = '/registration/[registrationId]/enrolment/completed',
   ENROLMENT_CANCELLED = '/registration/[registrationId]/enrolment/cancelled',
