+++ src/domain/enrolment/__tests__/utils.test.ts
@@ -1,10 +1,7 @@
 import { fakeRegistration, fakeSignup } from '../../../utils/mockDataUtils';
-import { registration } from '../../registration/__mocks__/registration';
 import { REGISTRATION_MANDATORY_FIELDS } from '../../registration/constants';
 import {
-  ATTENDEE_INITIAL_VALUES,
   ENROLMENT_FIELDS,
-  ENROLMENT_INITIAL_VALUES,
   NOTIFICATIONS,
   NOTIFICATION_TYPE,
 } from '../constants';
@@ -15,8 +12,6 @@ import {
   getEnrolmentDefaultInitialValues,
   getEnrolmentInitialValues,
   getEnrolmentNotificationsCode,
-  getEnrolmentNotificationTypes,
-  getEnrolmentPayload,
   isEnrolmentFieldRequired,
 } from '../utils';
 
@@ -44,109 +39,6 @@ describe('getEnrolmentNotificationsCode function', () => {
     expect(
       getEnrolmentNotificationsCode([NOTIFICATIONS.EMAIL, NOTIFICATIONS.SMS])
     ).toBe(NOTIFICATION_TYPE.SMS_EMAIL);
-  });
-});
-
-describe('getEnrolmentPayload function', () => {
-  const reservationCode = 'code';
-  it('should return single enrolment as payload', () => {
-    expect(
-      getEnrolmentPayload({
-        formValues: {
-          ...ENROLMENT_INITIAL_VALUES,
-          attendees: [ATTENDEE_INITIAL_VALUES],
-        },
-        registration,
-        reservationCode,
-      })
-    ).toEqual({
-      extra_info: '',
-      registration: registration.id,
-      reservation_code: 'code',
-      signups: [
-        {
-          city: '',
-          date_of_birth: null,
-          email: null,
-          extra_info: '',
-          first_name: '',
-          last_name: '',
-          membership_number: '',
-          native_language: null,
-          notifications: NOTIFICATION_TYPE.EMAIL,
-          phone_number: null,
-          responsible_for_group: true,
-          service_language: null,
-          street_address: null,
-          zipcode: null,
-        },
-      ],
-    });
-
-    const city = 'City',
-      dateOfBirth = '10.10.1999',
-      email = 'Email',
-      extraInfo = 'Extra info',
-      firstName = 'First name',
-      groupExtraInfo = 'Group extra info',
-      lastName = 'Last name',
-      membershipNumber = 'XXX-123',
-      nativeLanguage = 'fi',
-      notifications = [NOTIFICATIONS.EMAIL],
-      phoneNumber = '0441234567',
-      serviceLanguage = 'sv',
-      streetAddress = 'Street address',
-      zipcode = '00100';
-    const payload = getEnrolmentPayload({
-      formValues: {
-        ...ENROLMENT_INITIAL_VALUES,
-        attendees: [
-          {
-            city,
-            dateOfBirth,
-            extraInfo,
-            firstName,
-            inWaitingList: false,
-            lastName,
-            streetAddress,
-            zipcode,
-          },
-        ],
-        email,
-        extraInfo: groupExtraInfo,
-        membershipNumber,
-        nativeLanguage,
-        notifications,
-        phoneNumber,
-        serviceLanguage,
-      },
-      registration,
-      reservationCode,
-    });
-
-    expect(payload).toEqual({
-      extra_info: groupExtraInfo,
-      registration: registration.id,
-      reservation_code: reservationCode,
-      signups: [
-        {
-          city,
-          date_of_birth: '1999-10-10',
-          email,
-          extra_info: extraInfo,
-          first_name: firstName,
-          last_name: lastName,
-          membership_number: membershipNumber,
-          native_language: nativeLanguage,
-          notifications: NOTIFICATION_TYPE.EMAIL,
-          phone_number: phoneNumber,
-          responsible_for_group: true,
-          service_language: serviceLanguage,
-          street_address: streetAddress,
-          zipcode,
-        },
-      ],
-    });
   });
 });
 
@@ -303,25 +195,6 @@ describe('getEnrolmentInitialValues function', () => {
     expect(notifications).toEqual(expectedNotifications);
     expect(phoneNumber).toBe(expectedPhoneNumber);
     expect(serviceLanguage).toBe(expectedServiceLanguage);
-  });
-});
-
-describe('getEnrolmentNotificationTypes function', () => {
-  it('should return correct notification types', () => {
-    expect(
-      getEnrolmentNotificationTypes(NOTIFICATION_TYPE.NO_NOTIFICATION)
-    ).toEqual([]);
-    expect(getEnrolmentNotificationTypes(NOTIFICATION_TYPE.SMS)).toEqual([
-      NOTIFICATIONS.SMS,
-    ]);
-    expect(getEnrolmentNotificationTypes(NOTIFICATION_TYPE.EMAIL)).toEqual([
-      NOTIFICATIONS.EMAIL,
-    ]);
-    expect(getEnrolmentNotificationTypes(NOTIFICATION_TYPE.SMS_EMAIL)).toEqual([
-      NOTIFICATIONS.EMAIL,
-      NOTIFICATIONS.SMS,
-    ]);
-    expect(getEnrolmentNotificationTypes('lorem ipsum')).toEqual([]);
   });
 });
