+++ src/domain/enrolment/reservationTimer/__tests__/ReservationTimer.test.tsx
@@ -79,7 +79,7 @@ test('should show server errors when creating seats reservation fails', async ()
     )
   );
   singletonRouter.push({
-    pathname: ROUTES.CREATE_ENROLMENT,
+    pathname: ROUTES.CREATE_SIGNUP_GROUP,
     query: { registrationId: TEST_REGISTRATION_ID },
   });
   renderComponent({ showServerErrors });
@@ -100,7 +100,7 @@ test('should show modal if reserved seats are in waiting list', async () => {
     )
   );
   singletonRouter.push({
-    pathname: ROUTES.CREATE_ENROLMENT,
+    pathname: ROUTES.CREATE_SIGNUP_GROUP,
     query: { registrationId: TEST_REGISTRATION_ID },
   });
   renderComponent();
@@ -114,7 +114,7 @@ test('should show modal if reserved seats are in waiting list', async () => {
   await waitFor(() => expect(modal).not.toBeInTheDocument());
 });
 
-test('should route to create enrolment page if reservation is expired', async () => {
+test('should route to create signup group page if reservation is expired', async () => {
   const user = userEvent.setup();
 
   setEnrolmentFormSessionStorageValues({
@@ -123,7 +123,7 @@ test('should route to create enrolment page if reservation is expired', async ()
   });
 
   singletonRouter.push({
-    pathname: ROUTES.CREATE_ENROLMENT_SUMMARY,
+    pathname: ROUTES.CREATE_SIGNUP_GROUP_SUMMARY,
     query: { registrationId: registration.id },
   });
   renderComponent();
@@ -141,7 +141,7 @@ test('should route to create enrolment page if reservation is expired', async ()
 
   await waitFor(() =>
     expect(mockRouter.asPath).toBe(
-      `/registration/${registration.id}/enrolment/create`
+      `/registration/${registration.id}/signup-group/create`
     )
   );
 });
@@ -156,7 +156,7 @@ test('should reload page if reservation is expired and route is create enrolment
   });
 
   singletonRouter.push({
-    pathname: ROUTES.CREATE_ENROLMENT,
+    pathname: ROUTES.CREATE_SIGNUP_GROUP,
     query: { registrationId: registration.id },
   });
   renderComponent();
