+++ src/domain/enrolment/types.ts
@@ -70,31 +70,6 @@ export type Signup = {
   zipcode?: stringOrNull;
 };
 
-export type CreateSignupGroupMutationInput = {
-  extra_info: string;
-  registration: string;
-  reservation_code: string;
-  signups: SignupInput[];
-};
-
-export type CreateSignupGroupResponse = {
-  extra_info: string;
-  id: number;
-  registration: number;
-  signups: Signup[];
-};
-
-export type SignupGroup = {
-  created_at: stringOrNull;
-  created_by: stringOrNull;
-  extra_info: string;
-  id: number;
-  last_modified_at: stringOrNull;
-  last_modified_by: stringOrNull;
-  registration: number;
-  signups: Signup[];
-};
-
 export type DeleteEnrolmentMutationInput = {
   enrolmentId: string;
   registrationId: string;
