+++ src/domain/signupGroup/CreateSignupGroupPage.tsx
@@ -5,26 +5,26 @@ import LoadingSpinner from '../../common/components/loadingSpinner/LoadingSpinne
 import { ExtendedSession } from '../../types';
 import Container from '../app/layout/container/Container';
 import MainContent from '../app/layout/mainContent/MainContent';
+// eslint-disable-next-line max-len
+import AuthenticationRequiredNotification from '../enrolment/authenticationRequiredNotification/AuthenticationRequiredNotification';
+import EnrolmentForm from '../enrolment/enrolmentForm/EnrolmentForm';
+import { EnrolmentPageProvider } from '../enrolment/enrolmentPageContext/EnrolmentPageContext';
+import EnrolmentPageMeta from '../enrolment/enrolmentPageMeta/EnrolmentPageMeta';
+import { EnrolmentServerErrorsProvider } from '../enrolment/enrolmentServerErrorsContext/EnrolmentServerErrorsContext';
+import EventInfo from '../enrolment/eventInfo/EventInfo';
+import FormContainer from '../enrolment/formContainer/FormContainer';
+import useEventAndRegistrationData from '../enrolment/hooks/useEventAndRegistrationData';
+import { getEnrolmentDefaultInitialValues } from '../enrolment/utils';
 import { Event } from '../event/types';
 import NotFound from '../notFound/NotFound';
 import { Registration } from '../registration/types';
-// eslint-disable-next-line max-len
-import AuthenticationRequiredNotification from './authenticationRequiredNotification/AuthenticationRequiredNotification';
-import EnrolmentForm from './enrolmentForm/EnrolmentForm';
-import { EnrolmentPageProvider } from './enrolmentPageContext/EnrolmentPageContext';
-import EnrolmentPageMeta from './enrolmentPageMeta/EnrolmentPageMeta';
-import { EnrolmentServerErrorsProvider } from './enrolmentServerErrorsContext/EnrolmentServerErrorsContext';
-import EventInfo from './eventInfo/EventInfo';
-import FormContainer from './formContainer/FormContainer';
-import useEventAndRegistrationData from './hooks/useEventAndRegistrationData';
-import { getEnrolmentDefaultInitialValues } from './utils';
 
 type Props = {
   event: Event;
   registration: Registration;
 };
 
-const CreateEnrolmentPage: React.FC<Props> = ({ event, registration }) => {
+const CreateSignupGroupPage: React.FC<Props> = ({ event, registration }) => {
   const initialValues = getEnrolmentDefaultInitialValues();
   const { data: session } = useSession() as {
     data: ExtendedSession | null;
@@ -49,7 +49,7 @@ const CreateEnrolmentPage: React.FC<Props> = ({ event, registration }) => {
   );
 };
 
-const CreateEnrolmentPageWrapper: React.FC = () => {
+const CreateSignupGroupPageWrapper: React.FC = () => {
   const { event, isLoading, registration } = useEventAndRegistrationData();
 
   return (
@@ -57,7 +57,7 @@ const CreateEnrolmentPageWrapper: React.FC = () => {
       {registration && event ? (
         <EnrolmentPageProvider>
           <EnrolmentServerErrorsProvider>
-            <CreateEnrolmentPage event={event} registration={registration} />
+            <CreateSignupGroupPage event={event} registration={registration} />
           </EnrolmentServerErrorsProvider>
         </EnrolmentPageProvider>
       ) : (
@@ -67,4 +67,4 @@ const CreateEnrolmentPageWrapper: React.FC = () => {
   );
 };
 
-export default CreateEnrolmentPageWrapper;
+export default CreateSignupGroupPageWrapper;
