+++ src/domain/signupGroup/summaryPage/informantInfo/informantInfo.module.scss
