+++ src/domain/signupGroup/summaryPage/readOnlyTextInput/ReadOnlyTextInput.tsx
