+++ src/domain/signupGroup/summaryPage/readOnlyTextInput/readOnlyTextInput.module.scss
