+++ src/domain/signupGroup/summaryPage/signups/signup/signup.module.scss
