+++ src/domain/signupGroup/summaryPage/signups/signups.module.scss
@@ -1,3 +1,3 @@
-.attendees {
+.signups {
   margin-top: var(--spacing-l);
 }
