+++ src/domain/signupGroup/summaryPage/summaryEventInfo/summaryEventInfo.module.scss
