+++ src/domain/signupGroup/summaryPage/summaryPage.module.scss
